Add files via upload
The Alarm System project is an OpenCV-based application that detects motion through a camera and triggers an alarm. Using Python, the project leverages OpenCV to process video feeds and identify motion. When motion is detected, the system emits a beeping sound as an alert. This project demonstrates expertise in computer vision, real-time video processing, and integrating hardware for practical applications. It showcases skills in using OpenCV for motion detection, Python for scripting, and creating functional, real-world security solutions.